+++ src/containerapp/azext_containerapp/_help.py
@@ -727,18 +727,6 @@ helps['containerapp job create'] = """
                                     "queueLength": "5" "queueName": "foo" \\
               --scale-rule-auth "connection=my-connection-string-secret-name" \\
               --image imageName
-    - name: Create a container apps job hosted on a Connected Environment.
-      text: |
-          az containerapp job create -n MyContainerappsjob -g MyResourceGroup \\
-              --environment MyContainerappConnectedEnv
-              --environment-type connected
-              --trigger-type Manual \\
-              --replica-timeout 5 \\
-              --replica-retry-limit 2 \\
-              --replica-completion-count 1 \\
-              --parallelism 1 \\
-              --image imageName \\
-              --workload-profile-name my-wlp
 """
 
 helps['containerapp job update'] = """
@@ -1821,4 +1809,111 @@ helps['containerapp list'] = """
     - name: List container apps by environment type.
       text: |
           az containerapp list --environment-type connected
+"""
+
+# Connected Environment Commands
+helps['containerapp connected-env'] = """
+    type: group
+    short-summary: Commands to manage Container Apps Connected environments for use with Arc enabled Container Apps.
+"""
+
+helps['containerapp connected-env create'] = """
+    type: command
+    short-summary: Create a Container Apps connected environment.
+    long-summary: Create a Container Apps Connected environment for use with Arc enabled Container Apps.  Environments are an isolation boundary around a collection of container apps.
+    examples:
+    - name: Create a connected environment
+      text: |
+          az containerapp connected-env create -n MyContainerappConnectedEnv -g MyResourceGroup \\
+              --location eastus --custom-location MyCustomLocationResourceID
+"""
+
+helps['containerapp connected-env delete'] = """
+    type: command
+    short-summary: Delete a Container Apps connected environment.
+    examples:
+    - name: Delete a connected environment.
+      text: az containerapp connected-env delete -n MyContainerappConnectedEnv -g MyResourceGroup
+"""
+
+helps['containerapp connected-env show'] = """
+    type: command
+    short-summary: Show details of a Container Apps connected environment.
+    examples:
+    - name: Show the details of a connected environment.
+      text: |
+          az containerapp connected-env show -n MyContainerappConnectedEnv -g MyResourceGroup
+"""
+
+helps['containerapp connected-env list'] = """
+    type: command
+    short-summary: List Container Apps connected environments by subscription or resource group.
+    examples:
+    - name: List connected environments in the current subscription.
+      text: |
+          az containerapp connected-env list
+    - name: List connected environments by resource group.
+      text: |
+          az containerapp connected-env list -g MyResourceGroup
+"""
+
+# Container Apps Job Commands
+
+helps['containerapp job create'] = """
+    type: command
+    short-summary: Create a container apps job.
+    examples:
+    - name: Create a container apps job with Trigger Type as Manual.
+      text: |
+          az containerapp job create -n MyContainerappsjob -g MyResourceGroup \\
+              --environment MyContainerappEnv
+              --trigger-type Manual \\
+              --replica-timeout 5 \\
+              --replica-retry-limit 2 \\
+              --replica-completion-count 1 \\
+              --parallelism 1 \\
+              --image imageName \\
+              --workload-profile-name my-wlp
+    - name: Create a container apps job with Trigger Type as Schedule.
+      text: |
+          az containerapp job create -n MyContainerappsjob -g MyResourceGroup \\
+              --environment MyContainerappEnv
+              --trigger-type Schedule \\
+              --replica-timeout 5 \\
+              --replica-retry-limit 2 \\
+              --replica-completion-count 1 \\
+              --parallelism 1 \\
+              --cron-expression \"*/1 * * * *\" \\
+              --image imageName
+    - name: Create a container apps job with Trigger Type as Event.
+      text: |
+          az containerapp job create -n MyContainerappsjob -g MyResourceGroup \\
+              --environment MyContainerappEnv
+              --trigger-type Event \\
+              --replica-timeout 5 \\
+              --replica-retry-limit 2 \\
+              --replica-completion-count 1 \\
+              --parallelism 1 \\
+              --polling-interval 30 \\
+              --min-executions 0 \\
+              --max-executions 1 \\
+              --scale-rule-name queueJob \\
+              --scale-rule-type azure-queue \\
+              --scale-rule-metadata "accountName=mystorageaccountname" \\
+                                    "cloud=AzurePublicCloud" \\
+                                    "queueLength": "5" "queueName": "foo" \\
+              --scale-rule-auth "connection=my-connection-string-secret-name" \\
+              --image imageName
+    - name: Create a container apps job hosted on a Connected Environment.
+      text: |
+          az containerapp job create -n MyContainerappsjob -g MyResourceGroup \\
+              --environment MyContainerappConnectedEnv
+              --environment-type connected
+              --trigger-type Manual \\
+              --replica-timeout 5 \\
+              --replica-retry-limit 2 \\
+              --replica-completion-count 1 \\
+              --parallelism 1 \\
+              --image imageName \\
+              --workload-profile-name my-wlp
 """