+++ src/containerapp/azext_containerapp/commands.py
@@ -231,3 +231,9 @@ def load_command_table(self, _):
         g.custom_command('list', 'patch_list')
         g.custom_command('apply', 'patch_apply')
         g.custom_command('interactive', 'patch_interactive')
+
+    with self.command_group('containerapp connected-env', is_preview=True) as g:
+        g.custom_show_command('show', 'show_connected_environment')
+        g.custom_command('list', 'list_connected_environments')
+        g.custom_command('create', 'create_connected_environment', supports_no_wait=True, exception_handler=ex_handler_factory())
+        g.custom_command('delete', 'delete_connected_environment', supports_no_wait=True, confirmation=True, exception_handler=ex_handler_factory())