+++ src/spring/HISTORY.md
@@ -3,6 +3,7 @@ Release History
 Upcoming
 ---
 * Add arguments `--type` and `--git-sub-path` in `spring application-accelerator customized-accelerator create` and `spring application-accelerator customized-accelerator update` for accelerator fragment support.
+* Add new argument `--server-version` in `az spring app deploy and az spring app deployment create` to support war file deployment in Standard tier. 
 
 1.14.3
 ---