+++ src/spring/HISTORY.md
@@ -1,5 +1,13 @@
 Release History
 ===============
+Upcoming
+---
+* Add new argument `--server-version` in `az spring app deploy and az spring app deployment create` to support war file deployment in Standard tier. 
+
+1.14.3
+---
+* Make error message for `az spring app logs` more readable.  
+
 1.14.2
 ---
 * Add new command `az spring flush-virtualnetwork-dns-settings` to flush virtual network DNS settings for the service instance.